clang-tidy readability fixes.
Including
 - usage of static_cast instead of c-style cast,
 - single variable declaration per line,
 - removing const from return types
 - omit unnecessary string to char* conversions,

The used clang-tidy checks are:
google-readability-casting,
readability-const-return-type,
readability-container-size-empty,
readability-delete-null-pointer,
readability-deleted-default,
readability-else-after-return,
readability-function-size,
readability-identifier-naming,
readability-inconsistent-declaration-parameter-name,
readability-isolate-declaration,
readability-misleading-indentation,
readability-misplaced-array-index,
readability-named-parameter,
readability-redundant-*,
readability-simplify-*,
readability-static-accessed-through-instance,
readability-static-definition-in-anonymous-namespace,
readability-string-compare,
readability-uniqueptr-delete-release

+++ Tests/MathLib/TestEigenCSR.cpp
@@ -98,7 +98,8 @@ TEST(MathLibEigen, Eigen2CSR)
     int* col = mat.innerIndexPtr();
     double* data = mat.valuePtr();
 
-    for (int r=0; r<(int) nrows; ++r) {
+    for (int r = 0; r < nrows; ++r)
+    {
         EXPECT_EQ(ia[r], ptr[r]);
     }
